+++ src/Layer.h
@@ -25,7 +25,7 @@ public:
     }
     virtual MNC::Matrix feed_forward(MNC::Matrix &in) = 0;
     virtual MNC::Matrix back_propagation(const MNC::Matrix &in, const MNC::Matrix &inDer, const MNC::Matrix &err) = 0;
-    virtual void init(uint16_t inX, uint16_t inY, uint16_t inZ) = 0;
-    virtual void getOutDimensions(uint16_t &outX, uint16_t &outY, uint16_t &outZ) = 0;
+    virtual void init(uint32_t inX, uint32_t inY, uint32_t inZ) = 0;
+    virtual void getOutDimensions(uint32_t &outX, uint32_t &outY, uint32_t &outZ) = 0;
     
 };
